requirements.txt
bcrypt==4.0.1
cryptography
python-jose[cryptography]
fastapi
psycopg2-binary
PyMySQL
python-jose
python-multipart
SQLAlchemy
uvicorn
uvicorn[standard]
python-dotenv
passlib
pytest
httpx
pytest-asyncio
aiofiles
jinja2